tests/wpt/web-platform-tests/css/css-logical/parsing/margin-block-inline-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: margin-block and margin-inline sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-margin-block">
<meta name="assert" content="margin-block, margin-inline support the full grammar '<'margin-top'>{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('margin-block', '10px', {
  'margin-block-start': '10px',
  'margin-block-end': '10px'
});

test_shorthand_value('margin-block', '20% auto', {
  'margin-block-start': '20%',
  'margin-block-end': 'auto'
});

test_shorthand_value('margin-inline', '30%', {
  'margin-inline-start': '30%',
  'margin-inline-end': '30%'
});

test_shorthand_value('margin-inline', 'auto 40px', {
  'margin-inline-start': 'auto',
  'margin-inline-end': '40px'
});
</script>
</body>
</html>
